Add documentation on how to install "User Themes"
The "User Themes" GNOME Shell Extension is needed for this theme
to work from the ~/.themes folder. Without it, the ~/.themes folder
will not be registered.

This commit introduces a bit of explaining on the fact that this is
dependent upon the User Themes extension and how to enable it.