Use symbolic links so that git folder can be kept
When you keep the git folder in a specific place, you can get updates
from the repository just by doing a `git pull`. If you move the folder
from the git repository to the `~/.themes` folder, that mechanism is
lost.

This way, you keep the link to GitHub, allowing to import changes very
easily, without having to re-copy stuff.